fix setting an explicit wallet id when the wallet listing was paged
Summary:
Currently the tool gets a list of wallets when doing "bitgo wallet <id>" and checks for that id in the list, before actually doing the wallet get.
This breaks when the wallet listing is paged (user has access to too many wallets) because the walletid may not be in the first page.
This diff will always attempt a wallet get as long as an id is specified.

Reviewers: ben

Reviewed By: ben

Differential Revision: https://phabricator.bitgo.com/D1882