add fee rate calculation and psuedo-estimation
Summary:
SafeHD BCH recoveries are failing because we can't create new addresses on legacy migrated wallets.

We are creating new addresses because change outputs are being created when they shouldn't be.

We need to better estimate fees so that change outputs don't get created, and users don't run into this issue.

Reviewers: zacleids

Reviewed By: zacleids

Differential Revision: https://phabricator.bitgo.com/D7705